refactor: move atmosphere files into js/shapes/atmos/
Groups Atmosphere.js, AtmospherePrecompute.js, and atmosphere.test.js
under js/shapes/atmos/. Adds a one-line header to each explaining its
role (runtime rendering vs. LUT precompute vs. tests). Updates all
import paths in Planet.js, Star.js, and ThreeUI.js.
